% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIVPACS.new.R
\name{alt.RIVPACS.generic}
\alias{alt.RIVPACS.generic}
\title{River Invertebrate Prediction and Classification System (RIVPACS) Index calculation}
\usage{
alt.RIVPACS.generic(BenthicData, output_path, file_id)
}
\description{
The RIVPACS index is an Observed to Expected index comparing the infaunal community observed in a sample to that which would be expected to be in the
 location under reference conditions based on a predictive model based on latitude, longitude, and water depth
}
\details{
#   Details on the specifics of the calculation of the index can be found in Bay et al. 2021. Sediment Quality Assessment Techincal Support Manual. SCCWRP
}
