% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRI.R
\name{BRI.GenericOffshore.NH}
\alias{BRI.GenericOffshore.NH}
\title{Compute the benthic response index (BRI) score and BRI condition category. This version is the generic offshore one}
\usage{
BRI.GenericOffshore.NH(BenthicData)
}
\description{
The BRI is the abundance weighted pollution tolerance score of the organisms present in a benthic sample. The higher
  the BRI score, the more degraded the benthic community represented by the sample.
  This function should also work with offshore data
}
