% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRI.R
\name{BRI}
\alias{BRI}
\title{Compute the benthic response index (BRI) score and BRI condition category (generic version).}
\usage{
BRI(benthic_data)
}
\arguments{
\item{benthic_data}{a data frame containing benthic data and station information with at minimum:

   \strong{\code{stationid}} - an alpha-numeric identifier of the sampling location;

   \strong{\code{replicate}} - a numeric identifying the replicate number;

   \strong{\code{sampledate}} - the date of sample collection;

   \strong{\code{taxon}} - name of the organism. Use \code{NoOrganismsPresent} with 0 abundance for empty samples;

   \strong{\code{abundance}} - number of individuals counted;

   \strong{\code{exclude}} - "Yes" or "No" indicating if the taxon name is ambiguous.}

\item{logfile}{Path to a logfile. Default is an RMarkdown file in a timestamped logs directory.}

\item{verbose}{Logical. If TRUE, detailed logging output is produced. Default FALSE.}

\item{knitlog}{Logical. If TRUE, the log file is knitted to HTML upon completion. Default FALSE.}
}
\description{
The BRI is the 4th root abundance weighted pollution tolerance score of the organisms present in a sample
  relative to the 4th root abundance of all taxa in the sample with a tolerance score (aka p-code, p-value).
  The higher the BRI score, the more degraded the sample.
}
\details{
Details on the specifics of the calculation of the index can be found in Bay et al. 2021. Sediment Quality
  Assessment Technical Support Manual. SCCWRP Technical Report 777.

  Details on validation of the index can be found in Ranasinghe et al. 2009 Calibration and evaluation of five
  indicators of benthic community condition in two California bay and estuary habitats. Marine Pollution
  Bulletin 59:5-13.

  Background concepts of the index can be found in Smith et al 2001. Benthic Response Index for Assessing
  Infaunal Communities on the Southern California Mainland Shelf. Ecological Applications 11: 1073-1087.

  The BRI score is calculated as:
  \deqn{ BRI = \frac{\sum \left(\sqrt[4]{\textrm{Abundance}} \times P\right)}{\sum \sqrt[4]{\textrm{Abundance}}} }

  BRI condition categories for Southern California Marine Bays:
  \itemize{
    \item Reference: score < 39.96 (category score 1)
    \item Low Disturbance: 39.96 <= score < 49.15 (category score 2)
    \item Moderate Disturbance: 49.15 <= score < 73.27 (category score 3)
    \item High Disturbance: score >= 73.27 (category score 4)
  }

  This version uses the SQO Excel tool look-up list as decided by the Bight 23 index code subcommittee.
}
\examples{
\dontrun{
  BRI(my_benthic_data)
}

}
