% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{chemdata_prep}
\alias{chemdata_prep}
\title{Prepare raw chemistry data for analysis, per specifications in the}
\usage{
chemdata_prep(chemdata)
}
\arguments{
\item{chemdata}{a dataframe with the following columns:

   \code{StationID},

   \code{AnalyteName},

   \code{Result},

   \code{RL},

   \code{MDL}}
}
\description{
This function will not only calculate CSI and LRM, but it
will also get the overall integrated SQO Category and Score
for the stations that are given to it.
}
\details{
We make the assumption that Non-detects are marked with -88 in the result column. We also assume that
all constituents are expressed on a sediment dry-weight basis.
Specifically, all metals should be in mg/dry kg and all organic constituents should be in ug/dry kg.
}
\examples{
data(chem_sampledata) # load sample data to your environment
chemdata_prep(chem_sampledata) # get scores and see output

}
