% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logutils.R
\name{create_download_link}
\alias{create_download_link}
\title{Create download link within an RMarkdown log file}
\usage{
create_download_link(
  data,
  logfile,
  filename,
  linktext = "Download the data",
  include.row.names = FALSE,
  verbose = F
)
}
\arguments{
\item{data}{data to be downloaded}

\item{logfile}{the main log file filepath}

\item{filename}{name of the csv file to be downloaded}

\item{linktext}{text on the download link}

\item{include.rownames}{should rownames of the dataframe be included in the CSV file?}
}
\description{
Create download link within an RMarkdown log file
}
