% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ed.14.ptaxa}
\alias{ed.14.ptaxa}
\title{Edition 14 taxon-to-p-code lookup for offshore BRI}
\format{
A data frame with 1196 rows and 2 columns:
\describe{
  \item{p_code}{Integer p-code identifying the tolerance score group}
  \item{taxon}{SCAMIT Edition 14 taxon name}
}
}
\source{
SCCWRP internal reference data. Stored in \code{data/pcode_14.RData}.
}
\usage{
ed.14.ptaxa
}
\description{
Maps SCAMIT Edition 14 taxon names to p-codes used in the offshore Benthic
Response Index (BRI) calculation. Each p-code links to depth-zone tolerance
scores in \code{\link{pcodes}}.
}
\seealso{
\code{\link{pcodes}}, \code{\link{BRI.Offshore}}
}
\keyword{datasets}
