% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ed14.rollups}
\alias{ed14.rollups}
\title{Edition 14 taxon rollup lookup for SQO retrofitting}
\format{
A data frame with 988 rows and 3 columns:
\describe{
  \item{sqo.name}{SQO-compatible parent taxon name}
  \item{ed.14_daughters}{SCAMIT Edition 14 taxon name that rolls up to the SQO parent}
  \item{join_level}{Taxonomic level at which the rollup join is applied}
}
}
\source{
SCCWRP internal reference data. Stored in \code{data/ed14_to_SQO_v2.RData}.
}
\usage{
ed14.rollups
}
\description{
Maps Edition 14 daughter taxa back to their parent SQO-compatible names.
Used by \code{\link{benthicdata_prep}} to aggregate modern taxonomy into
the coarser SQO name space before index calculations.
}
\seealso{
\code{\link{ed.14.complex}}, \code{\link{SoCal.SQO.xls.ed14.link}}, \code{\link{benthicdata_prep}}
}
\keyword{datasets}
