% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logutils.R
\name{init.log}
\alias{init.log}
\title{Function to write to a log file}
\usage{
init.log(
  logfile,
  base.func.name,
  type = "text",
  current.time = Sys.time(),
  is.base.func = T,
  verbose = F,
  title = "Log",
  libraries = c("rmarkdown")
)
}
\arguments{
\item{logfile}{Path to the file where the content will be written}

\item{base.func.name}{the function name where the call originated - to be written at the top of the log file}

\item{current.time}{time the function got called}

\item{is.base.func}{log file shouldnt get initialized if the function gets called within a function that is not the originating one
Just tells us whether to actually create the log file or not}

\item{verbose}{This is what it will be called in parent functions - verbose tells whether to actually write to the logfile or not
This will allow us to write the code in the other functions and files without a bunch of if statements .}
}
\description{
Function to write to a log file
}
\examples{
init.log(logfile, sys.call())
}
