% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knit.log}
\alias{knit.log}
\title{Knit an R Markdown log file to HTML (internal)}
\usage{
knit.log(logfile, verbose = F, knitlog = F)
}
\arguments{
\item{logfile}{path to the .Rmd log to render}

\item{verbose}{only render when TRUE}

\item{knitlog}{only render when TRUE}
}
\description{
No-op unless both verbose and knitlog are TRUE and the logfile is a .Rmd. Renders to a
sibling .html. Centralizes the render step used by the SQO logging functions.
}
