% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{LRM}
\alias{LRM}
\title{Calulate the Indices for Chemistry Line of Evidence to determine Station Assessment
in other words, how impacted a site is}
\usage{
data(lrm_table)

data(csi_weight)
}
\arguments{
\item{chemdata}{a dataframe with the following headings
\code{StationID} - an alpha-numeric identifier of the location;
\code{AnalyteName} - Name of the Chemical Compound
\code{Result} - the concentration of the Analyte at that station
\code{RL} - Reporting Limit
\code{MDL} Method Detection Limit}
}
\description{
This is done by getting the :
  LRM (Logistic Regression Model) Score
  CSI (Chemical Score Index)
  The Average of the Above two Scores, which is hard to come up with a name for
}
\details{
The ultimate guide for this thing is the CASQO Technical Manual Chapter 3 (page 18 to 36)
}
