% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SoCal.SQO.xls.ed14.link}
\alias{SoCal.SQO.xls.ed14.link}
\title{SQO Excel tool Edition 14 taxon link table}
\format{
A data frame with 1218 rows and 9 columns:
\describe{
  \item{original_sqo_taxon}{Original SQO-compatible taxon name}
  \item{change}{Type of taxonomic change applied}
  \item{type}{Change category (e.g. "one-to-one", "complex", "rollup")}
  \item{ed_14_taxon}{Corresponding SCAMIT Edition 14 taxon name}
  \item{ibi_sensitive_taxa}{IBI sensitivity designation ("S" = sensitive)}
  \item{mollusc_designation}{Mollusc flag ("Mollusc" or blank)}
  \item{crustacean_designation}{Crustacean flag}
  \item{bri_tolerance_score}{BRI tolerance score for the taxon}
  \item{used_in_rivpacs}{Whether the taxon is included in the RIVPACS model}
}
}
\source{
SCCWRP internal reference data. Stored in \code{data/ed14_to_SQO_v2.RData}.
}
\usage{
SoCal.SQO.xls.ed14.link
}
\description{
Master lookup table linking original SQO taxon names to SCAMIT Edition 14
equivalents, with index-specific designations (IBI sensitivity, mollusc/crustacean
flags, BRI tolerance score, RIVPACS inclusion). This is the primary reference
used by \code{\link{benthicdata_prep}} to retrofit modern taxonomy.
}
\seealso{
\code{\link{ed.14.complex}}, \code{\link{ed14.rollups}}, \code{\link{benthicdata_prep}}
}
\keyword{datasets}
