% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIVPACS.R
\name{SoCalRivpacs}
\alias{SoCalRivpacs}
\title{This is the original RIVPACS function. Called by the function called "RIVPACS" which outputs the data in a more friendly format.
This function does most of the heavy lifting though
Exporting it so it can be more accessible}
\usage{
SoCalRivpacs(
  Pcutoff = 0.5,
  reference.groups = socal.reference.groups,
  observed.predictors = socal.example.habitat,
  reference.taxa = socal.reference.taxa,
  group.means = socal.reference.group.means,
  reference.cov = socal.reference.covariance,
  observed.taxa = socal.example.taxa,
  logfile = file.path(getwd(), "logs", format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M-\%S"),
    "log.txt"),
  verbose = F
)
}
\description{
This is the original RIVPACS function. Called by the function called "RIVPACS" which outputs the data in a more friendly format.
This function does most of the heavy lifting though
Exporting it so it can be more accessible
}
