% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQOUnified.R
\name{SQOUnified}
\alias{SQOUnified}
\title{Compute the SQO index scores.}
\usage{
data(DB)
}
\arguments{
\item{DB}{A data file string name that we want to compute SQO scores for.}

\item{SQO}{A list of the type of SQO scores that we want to compute
(e.g., \code{c("MAMBI", "RBI")}).
The default is \code{"all"}, meaning that all scores will be computed.}
}
\description{
Compute the SQO index scores.
}
\examples{
SQOUnified(DB = benthic_data, SQO = "all")
SQOUnified(DB = benthic_data, SQO = "MAMBI")
SQOUnified(DB = benthic_data, SQO = c("MAMBI", "RBI")) --> To be implemented in
the future
}
