% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TidalFresh_Standards}
\alias{TidalFresh_Standards}
\title{M-AMBI reference standards for tidal-fresh stations}
\format{
A data frame with 2 rows and 7 columns:
\describe{
  \item{stationid}{Station identifier}
  \item{replicate}{Replicate number}
  \item{sampledate}{Sample collection date}
  \item{ambi_score}{AMBI score at the reference station}
  \item{H}{Shannon diversity (H') at the reference station}
  \item{oligo_pct}{Percentage oligochaetes at the reference station}
  \item{SalZone}{Salinity zone designation}
}
}
\source{
SCCWRP internal reference data. Stored in \code{data/TidalFresh_Standards.RData}.
}
\usage{
TidalFresh_Standards
}
\description{
Reference condition values for the M-AMBI index at tidal-freshwater stations
in Southern California. Used by \code{\link{MAMBI}} alongside
\code{\link{Saline_Standards}} for index normalisation.
}
\seealso{
\code{\link{Saline_Standards}}, \code{\link{MAMBI}}
}
\keyword{datasets}
