% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity.R
\name{tox.summary}
\alias{tox.summary}
\title{Get the Statistical Summary for Tox Results Data}
\usage{
tox.summary(toxresults)
}
\arguments{
\item{toxresults}{a dataframe with the following columns: stationid, toxbatch, species, sampletypecode
matrix, labrep, result. This data must also include the control samples
(stationcode 0000, sampletypecode CNEG etc)}
}
\description{
This function will calculate the tox summary,

given an argument which is a dataframe of tox results data
}
\details{
The input data here is called toxresults, and it is structured as follows:

\strong{toxresults} -  a dataframe that contains the toxicity results

   \code{stationid} - an alpha-numeric identifier of the location;

   \code{toxbatch} - the toxbatch id - used to join with the control sample

   \code{species} - The Genus and species of the animale that was tested

   \code{sampletypecode} - The sampletype used Grab, CNEG etc. Control samples must be included

   \code{matrix} - Whole Sediment, Sediment Water Interface, etc. Probably useless to include.
                   I Just have it to make sure they dont put Reference Toxicant

   \code{labrep} - There should be 5 per station, species pair

   \code{result} - the percentage that survived the test, or had normal development
}
\examples{
data(tox_sampledata)
tox.summary(tox_sampledata)

}
