% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us.mambi.eg.values.04_23_24}
\alias{us.mambi.eg.values.04_23_24}
\title{US M-AMBI ecological group reference values (April 2023 update)}
\format{
A data frame with 7185 rows and 10 columns:
\describe{
  \item{Taxon}{Taxon name}
  \item{Exclude}{Whether the taxon is excluded from index calculations}
  \item{Hybrid}{Whether the taxon is a hybrid designation}
  \item{US}{Ecological group assignment for US waters (overall)}
  \item{Standard}{Standard EG value}
  \item{US_East}{EG assignment for US East Coast}
  \item{US_Gulf}{EG assignment for US Gulf Coast}
  \item{US_West}{EG assignment for US West Coast}
  \item{Oligochaeta}{Whether the taxon is an oligochaete (relevant for tidal-fresh)}
  \item{note}{Additional notes on the EG assignment}
}
}
\source{
SCCWRP internal reference data. Stored in \code{data/us mambi egvalues 04-23-24.RData}.
}
\usage{
us.mambi.eg.values.04_23_24
}
\description{
Ecological group (EG) assignments and reference values for taxa used in the
US M-AMBI calculation. Covers taxa from US marine, estuarine, and tidal-fresh
habitats. Updated April 2023.
}
\seealso{
\code{\link{Saline_Standards}}, \code{\link{TidalFresh_Standards}}, \code{\link{MAMBI}}
}
\keyword{datasets}
