% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthic.R
\name{benthic.sqo}
\alias{benthic.sqo}
\title{Get Benthic Related Indices, and the Integrated Benthic SQO Score}
\usage{
benthic.sqo(benthic_data)
}
\arguments{
\item{benthic_data}{a data frame. This data frame must contain the following
 information with these headings:

   \code{StationID} - an alpha-numeric identifier of the location;

   \code{Replicate} - a numeric identifying the replicate number of samples taken at the location;

   \code{SampleDate} - the date of sample collection;

   \code{Latitude} - latitude in decimal degrees;

   \code{Longitude} - longitude in decimal degrees. Make sure there is a negative sign for the Western coordinates;

   \code{Taxon} - name of the fauna, ideally in SCAMIT ed12 format, do not use sp. or spp.,
       use sp only or just the Genus. If no animals were present in the sample use
       NoOrganismsPresent with 0 abundance;

   \code{Abundance} - the number of each Species observed in a sample;

   \code{Salinity} - the salinity observed at the location in PSU, ideally at time of sampling;

   \code{Stratum} - The stratum under which the station falls (Bays, Estuaries, etc);

   \code{Exclude} - Yes or No;}
}
\description{
This function will calculate, RBI, IBI, BRI, RIVPACS, MAMBI,
as well as the Integrated Benthic SQO score / category.
The ultimate guide for these indices (EXCEPT MAMBI) is the CASQO Technical Manual
pages 64 to 94
}
\examples{
data(benthic_sampledata) # load the sample data
benthic.sqo(benthic_sampledata) # get scores and see output

}
