% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRI.R
\name{BRI}
\alias{BRI}
\title{Compute the benthic response index (BRI) score and BRI condition category.}
\usage{
BRI(benthic_data)
}
\arguments{
\item{BenthicData}{a data frame with the following headings

   \strong{\code{StationID}} - an alpha-numeric identifier of the location;

   \strong{\code{Replicate}} - a numeric identifying the replicate number of samples taken at the location;

   \strong{\code{SampleDate}} - the date of sample collection;

   \strong{\code{Latitude}} - latitude in decimal degrees;

   \strong{\code{Longitude}} - longitude in decimal degrees.
   Make sure there is a negative sign for the Western coordinates;

   \strong{\code{Species}} - name of the fauna, ideally in SCAMIT ed12 format, do not use sp. or spp.,
       use sp only or just the Genus. If no animals were present in the sample use
       NoOrganismsPresent with 0 abundance;}
}
\description{
The BRI is the abundance weighted pollution tolerance score of the organisms present in a benthic sample. The higher
  the BRI score, the more degraded the benthic community represented by the sample.
}
\details{
The BRI is the abundance weighted pollution tolerance score of the organisms present in a benthic sample. The higher
  the BRI score, the more degraded the benthic community represented by the sample.

  Two types of data are needed to calculate the BRI:

  (1) the abundance of each species and
  (2) its pollution tolerance score, P.

  The P values are available for most species present in the assemblage. Only species for which P values are avialable
  are used in the BRI calculations. P values showld be obtained for the appropriate habitat and from the most
  up-to-date list available.

  The first step in the BRI calculation is to compute the 4th root of the abundance of each taxon in the sample for
  which P values are available. The next step is to multiply the 4th root abundance value by the P value, for each
  taxon.

  Next, separately sum all of the 4th roots of the abundances and all of the products of the 4th roots of abundance
  and P values. Taxa that lack P values are not included in either sum. The next step is to calculate the BRI score
  as:

  \deqn{ \frac{\sum \left(\sqrt[p]{\textrm{Abundance}} \right) \times P}{\sum \sqrt[p]{\textrm{Abundance}}} }

  The last step is to compare the BRI score to the BRI threshold values in Table 5 to determine the BRI category and
  category score.

  <Table 5. To be included in R markdown file>
}
\examples{
data(benthic_sampledata) # load sample data
BRI(benthic_sampledata) # see the output

}
