% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{chem.sqo}
\alias{chem.sqo}
\title{Get Integrated Chem SQO Scores and Categories}
\usage{
chem.sqo(chemdata)
}
\arguments{
\item{chemdata}{a dataframe with the following columns:

   \code{StationID},

   \code{AnalyteName},

   \code{Result},

   \code{RL},

   \code{MDL}}
}
\description{
This function will not only calculate CSI and LRM, but it
will also get the overall integrated SQO Category and Score
for the stations that are given to it.
}
\examples{
data(chem_sampledata) # load sample data to your environment
chem.sqo(chem_sampledata) # get scores and see output

}
