% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{chem.sqo}
\alias{chem.sqo}
\title{Get Integrated Chem SQO Scores and Categories}
\usage{
chem.sqo(chemdata)
}
\arguments{
\item{chemdata}{a dataframe with the following columns:

   \code{StationID}

   \code{AnalyteName}

   \code{Result}

   \code{RL}

   \code{MDL}

   \code{units} (optional) -  Metals should be in mg/dry kg (mg/kg dw) and all organic constituents should be in ug/dry kg (ug/kg dw).

   \code{fieldrep} (optional) - data will be filtered to where fieldrep = 1

   \code{labrep} (optional) - data will be filtered to where labrep = 1

   \code{sampletypecode} (optional) - data will be filtered to where sampletypecode = Result (to avoid including data from QA/QC samples)}
}
\description{
This function will not only calculate CSI and LRM, but it
will also get the overall integrated SQO Category and Score
for the stations that are given to it.
}
\examples{
data(chem_sampledata) # load sample data to your environment
chem.sqo(chem_sampledata) # get scores and see output

}
