% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{CSI}
\alias{CSI}
\title{Chemical Score Index}
\usage{
CSI(chemdata)
}
\arguments{
\item{chemdata}{a dataframe with the following headings:

   \code{StationID}

   \code{AnalyteName}

   \code{Result}

   \code{RL}

   \code{MDL}

   \code{fieldrep} (optional) - data will be filtered to where fieldrep = 1

   \code{labrep} (optional) - data will be filtered to where labrep = 1

   \code{sampletypecode} (optional) - data will be filtered to where sampletypecode = Result (to avoid including data from QA/QC samples)}
}
\description{
This function calculates the Chemical Score Index for the given sites.
The ultimate guide for this function is the CASQO Technical Manual Chapter 3
(beginning of page 34 to the beginning of page 36)
}
\examples{
data(chem_sampledata) # load sample data to your environment
CSI(chem_sampledata) # get scores and see output

}
