% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry.R
\name{LRM}
\alias{LRM}
\title{Logistic Regression Model Score}
\usage{
LRM(chemdata)
}
\arguments{
\item{chemdata}{a dataframe with the following headings:

   \code{StationID}

   \code{AnalyteName}

   \code{Result}

   \code{RL}

   \code{MDL}

   \code{units} (optional) -  Metals should be in mg/dry kg (mg/kg dw) and all organic constituents should be in ug/dry kg (ug/kg dw).

   \code{fieldrep} (optional) - data will be filtered to where fieldrep = 1

   \code{labrep} (optional) - data will be filtered to where labrep = 1

   \code{sampletypecode} (optional) - data will be filtered to where sampletypecode = Result (to avoid including data from QA/QC samples)}
}
\description{
This function calculates the Logistic Regression Model score
in order to calculate the Chemistry SQO scores for the given sites.
The ultimate guide for this function is the CASQO Technical Manual Chapter 3
(end of page 31 to beginning of page 34)
}
\examples{
data(chem_sampledata) # load sample data to your environment
LRM(chem_sampledata) # get scores and see output

}
