% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMBI.R
\name{MAMBI}
\alias{MAMBI}
\title{Compute the multivariate AMBI (M-AMBI) index score.}
\usage{
data(benthic_data)

data(EG_Ref)

data(Saline_Standards)

data(TidalFresh_Standards)
}
\arguments{
\item{BenthicData}{a data frame with the following information with these headings:
\code{StationID} - an alpha-numeric identifier of the location;
\code{Replicat} - a numeric identifying the replicate number of samples taken at the location;
\code{SampleDate} - the date of sample collection;
\code{Latitude} - latitude in decimal degrees;
\code{Longitude} - longitude in decimal degrees. Make sure there is a negative sign for the Western coordinates;
\code{Species} - name of the fauna, ideally in SCAMIT ed12 format, do not use sp. or spp.,
    use sp only or just the Genus. If no animals were present in the sample use
    NoOrganismsPresent with 0 abundance;
\code{Abundance} - the number of each Species observed in a sample;
\code{Salinity} - the salinity observed at the location in PSU, ideally at time of sampling.}

\item{EG_File_Name}{A quoted string with the name of the csv file with the suite of US Ecological Groups assigned
initially in Gillett et al. 2015. This EG file has multiple versions of the EG values and a Yes/No designation
if the fauna are Oligochaetes or not. The default file is the Ref - EG Values 2018.csv file included with this
code. Replace with other files as you see fit, but make sure the file you use is in a similar format and uses
the same column names. Additionally, new taxa can be added at the bottom of the list with the EG values the user
feels appropriate, THOUGH THIS IS NOT RECOMMENDED}

\item{EG_Scheme}{A quoted string with the name of the EG Scheme to be used in the AMBI scoring. The default is
Hybrid, though one could use US (all coasts), Standard (Values from Angel Borja and colleagues),
US_East (US East Coast), US_Gulf (US Gulf of Mexico Coast), or US_West (US West Coast).}
}
\description{
Compute the multivariate AMBI (M-AMBI) index score.
}
\examples{
MAMBI.DJG.alt(benthic_data, EG_File_Name="data/Ref - EG Values 2018.csv", EG_Scheme="Hybrid")
MAMBI.DJG.alt(benthic_data, EG_File_Name="data/Ref - EG Values 2018.csv", EG_Scheme="US_Gulf")
}
