% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQOUnified.R
\name{SQOUnified}
\alias{SQOUnified}
\title{Compute the SQO index scores.
Based on the Technical Manual - June 2021 edition
https://ftp.sccwrp.org/pub/download/DOCUMENTS/TechnicalReports/777_CASQO_TechnicalManual.pdf}
\usage{
data(benthic_data)
data(site_assessment_criteria)
}
\arguments{
\item{benthic_data}{A data file string name that we want to compute SQO scores for.}

\item{SQO}{A list of the type of SQO scores that we want to compute
(e.g., \code{c("MAMBI", "RBI")}).
The default is \code{"all"}, meaning that all scores will be computed.}
}
\description{
Compute the SQO index scores.
Based on the Technical Manual - June 2021 edition
https://ftp.sccwrp.org/pub/download/DOCUMENTS/TechnicalReports/777_CASQO_TechnicalManual.pdf
}
\examples{
SQOUnified(benthic_data = benthic_data, LOE = "all")
SQOUnified(benthic_data = benthic_data, LOW = 'tox')

}
