% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity.R
\name{tox.summary}
\alias{tox.summary}
\title{Get the Statistical Summary for Tox Results Data}
\usage{
tox.summary(toxresults)
}
\arguments{
\item{toxresults}{a dataframe with the following columns: stationid, toxbatch, species, sampletypecode
   matrix, labrep, result. This data must also include the control samples
   (stationcode 0000, sampletypecode CNEG etc)

   The input dataframe is structured as follows

   \strong{\code{lab}} -  (optional) The laboratory which performed the test

   \strong{\code{stationid}} - an alpha-numeric identifier of the location;

   \strong{\code{toxbatch}} - the toxbatch id - used to join with the control sample

   \strong{\code{species}} - The Genus and species of the animale that was tested

   \strong{\code{sampletypecode}} - The sampletype used Grab, CNEG etc. Control samples must be included

   \strong{\code{matrix}} - (optional) Whole Sediment, Sediment Water Interface, etc. Be sure to not include Reference Toxicants

   \strong{\code{fieldrep}} - Field replicate

   \strong{\code{labrep}} - There should be 5 per station, species pair

   \strong{\code{result}} - the percentage that survived the test, or had normal development}
}
\description{
This function will calculate the tox summary,
given an argument which is a dataframe of tox results data.
By default, control samples will only be those with sampletypecode of CNEG. You may allow for others, such as CNSL.
A warning will be issued if multiple controls are found within one batch
}
\examples{
data(tox_sampledata)
tox.summary(tox_sampledata)

}
